requirements.txt
python-dotenv
requests
pydub
ffmpeg